Moved HtmlUtils to helpers
