+++ src/helpers/favicon.ts
@@ -1,4 +1,4 @@
-import { HtmlUtils } from '../HtmlUtils';
+import { HtmlUtils } from './HtmlUtils';
 import { Debug } from '../Debug';
 import { safeFetch } from './safeFetch';
 import * as urlUtils from './urlUtils';
